[Pal/VM,TDX] Use lazy memory allocation for anon memory
Previously, VM/TDX PALs had all memory pages present (P bit = 1) all the
time, and they did a memset-to-zero of memory regions as soon as they
were requested in `PalVirtualMemoryAlloc()`. This commit changes it to
lazy memory allocation for anonymous memory: pages are marked as not
present (P bit = 0) and memset-to-zero on demand, in the #PF exception
handler. This is purely a performance optimization. Same for file-backed
memory.

As a side-effect of this change, the legacy "kernel stack pointer" in
TSS is replaced with the newer "interrupt stack table" entry. This is
because the "kernel stack pointer" changes to the dedicated stack
`_sys_interrupt_stack` only on ring 3 -> ring 0 transitions, so if e.g.
a #PF would happen while in ring 0, the interrupt would be processed on
the same stack, and the pages of this "normal" stack could be not
present in PTEs (leading to a double-fault: #PF during #PF handling). In
contrast, IST entries switch the stack unconditionally.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>