[Pal/VM] Use well-defined initial stack
Previously, the stack started (to grow downwards) from the base address
of the PAL binary (which is at 1.5MB or 0x180000). This stack space was
not accounted for and could lead to memory corruptions (if e.g. the PAL
would decide to allocate some objects there via _PalVirtualMemoryAlloc).
Now the stack is part of the PAL binary and has a size of 16KB.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>