[PAL/common-vm] vsock: use proper queue of pending connections
Previously, our vsock implementation allowed a single pending connection
on a listening socket (i.e., a queue of size 1). This breaks
applications like the `memtier_benchmark` client of Memcached, which
assume that they can create many client connections simultaneously.

This commit changes the previous naive implementation to a proper queue
of pending connections; actually it's a simple static array of size 128
(a ring buffer with all operations O(1), so it is efficient).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>