[PAL/vm-common] vsock: process TX even if processing RX failed
Previously, the vsock bottomhalves thread exited if during receiving
packets on the RX queue an error happened. Thus, processing on the TX
queue (cleaning up slots of used-by-host packets) didn't happen until
the RX queue went back into "normal" state.

However, the RX queue can be full, which results in the NOMEM error.
This error is benign -- it will disappear after some time, when the app
reads the packets. So there is no legit reason to *not* continue with
processing on the TX queue.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>