[PAL/vm-common] Increase limit on size of cmdline args
Previously, the buffer to read cmdline args was allocated on the stack
and thus had allowed for only 512 chars. This turns out to be not enough
for workloads with many args, e.g., TensorFlow.

This commit increases the limit on the size of cmdline args to 4096
chars. Since it's dangereous to allocate 4KB on the stack, we change the
buffer from the stack to being a global variable.

Signed-off-by: Dimitrios Stavrakakis <dimitrios.stavrakakis@intel.com>