[PAL/vm-common] console: write stdout/stderr in chunks
Previously, PAL devices tried to write the whole buffer of the app into
the virtio-console shared buffer. However, the latter is only 4KB in
size, and an app buffer of e.g. 8KB resulted in a NOMEM error. This was
detected on a Python microbenchmark that constantly prints to stdout.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>