[PAL/vm-common] vsock: correctly implement `shutdown()`
Previously, the `shutdown()` operation (i.e., `VIRTIO_VSOCK_OP_SHUTDOWN`
vsock operation) was incorrectly emulated as the `close()` operation. In
particular, our code didn't take into account the RCV/SEND/COMPLETE flag
of the SHUTDOWN operation. Also, the SHUTDOWN operation moved the
connection state to "closed", which prevented any further operations
like read/write on this connection, even if the shutdown was only
partial. Also, a benign RST reply on the SHUTDOWN operation wasn't
ignored (as it should be), but instead moved the connection state to
"closed".

This commit fixes all the above issues. Now we keep the connection
opened in case of a partial shutdown, and we don't panic on benign RST
packets. This fixes workloads like lighttpd + wrk.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>