[PAL/TDX] Unset shared bit in the resulting addr of MapGPA
In TDX, TDG.VP.VMCALL<MapGPA> guest-host interface may not map the whole
requested GPA range. In this case, this interface returns
TDG.VP.VMCALL_RETRY error code and sets register R11 (`failed_addr`) to
the first GPA at which the mapping failed.

Our TDX PAL sanity checks that this returned `failed_addr` falls into
the original requested GPA range. However, there was a bug that TDX
returns this failing address as-is (in our case, with the shared bit
set), so a naive check is out-of-range (because we compare against
no-shared-bit addresses). This commit unsets the shared bit of the
failing address before performing the sanity check.

This bug was detected on Ubuntu 24.04 with Linux 6.8.0 and QEMU v8.2.1.
Apparently this version of QEMU/KVM performs MapGPA only in 64MB chunks,
and if the requested GPA range is larger, errors our with VMCALL_RETRY.
Previous versions didn't exhibit this behavior, so the bug went
unnoticed for a long time.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>