[LibOS] Rewire RIP of to-save and to-restore contexts for VM PALs
Previously, the LibOS logic had two places that assumed ring-3
environment and did not account for the need to execute `sysret`:
- Preparing sigframes for the application signal handler; the
  interrupted app context is saved in the sigframe.
- Performing `rt_sigreturn()` syscall after the application signal
  handler is done; the app context is restored from the previously-saved
  sigframe.

In case of VM-based PALs, where LibOS runs in ring-0 and app runs in
ring-3, there is a wrapper around syscall enter/exit, see `syscall_asm`
and `sysret_asm` in vm-common/kernel_events.S file. This wrapper rewires
the context RIP: upon syscall entry, it saves the app context RIP into a
TCB-local variable and sets RIP to the address of the `sysret_asm`
routine, and upon syscall exit, it restores app context RIP from the
TCB-local variable.

This wrapper rewiring was not accounted for in the LibOS logic, meaning
that the LibOS would prepare the sigframe with RIP not of the app
context but of the `sysret_asm` routine. Similarly, the LibOS would
restore not the app context's RIP but the RIP of `sysret_asm`. This
commit adds VM PAL-specific RIP fixups to the LibOS logic.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>